+++ scripts/build-demo.js
@@ -6,6 +6,7 @@
 
 'use strict';
 
+const fs = require('fs');
 const utils = require('./utils');
 const paths = require('./paths');
 const configFactory = require('./webpack.config-demo');